requirement.txt
mysql-connector-python
requests
datetime
matplotlib